📖 docs: document public members
