+++ analysis_options.yaml
@@ -138,7 +138,7 @@ linter:
     - prefer_typing_uninitialized_variables
     - prefer_void_to_null
     - provide_deprecation_message
-    # - public_member_api_docs
+    - public_member_api_docs
     - recursive_getters
     - require_trailing_commas
     - sized_box_for_whitespace
