+++ lib/src/core/data/models/attribute.dart
@@ -2,24 +2,25 @@ import 'dart:math';
 
 import 'package:meta/meta.dart';
 
+/// Character's trait interface.
 @immutable
 abstract class Attribute {
+  /// Initialize [points]
   const Attribute(this.points);
 
+  /// Current value of this attribute.
+  /// Must be a positive number.
   final int points;
 
+  /// Clone attribute with different value.
   Attribute copyWith({int? points});
 
+  /// Text representation for this attribute.
   String label();
 
+  /// Compute numbers of skills the character must substract for upgrading this
+  /// attribute by one.
   int get skillsCost => max(1, (points / 5).ceil());
-
-  int operator -(int by) => points - by;
-  int operator +(int by) => points + by;
-  bool operator >(Attribute other) => points > other.points;
-  bool operator >=(Attribute other) => points >= other.points;
-  bool operator <(Attribute other) => points < other.points;
-  bool operator <=(Attribute other) => points <= other.points;
 
   @override
   bool operator ==(Object other) {
@@ -32,7 +33,9 @@ abstract class Attribute {
   int get hashCode => points.hashCode;
 }
 
+/// Character's health trait. Will be regenerated after each fight.
 class Health extends Attribute {
+  /// Initialize health [points]
   const Health(int points) : super(points);
 
   @override
@@ -48,7 +51,9 @@ class Health extends Attribute {
   String toString() => 'Health($points)';
 }
 
+/// A character's trait to compute the damage during a fight.
 class Attack extends Attribute {
+  /// Initialize attack [points]
   const Attack(int points) : super(points);
 
   @override
@@ -61,7 +66,9 @@ class Attack extends Attribute {
   String toString() => 'Attack($points)';
 }
 
+/// A character's trait to reduce the damage during a fight.
 class Defense extends Attribute {
+  /// Initialize defense [points]
   const Defense(int points) : super(points);
 
   @override
@@ -74,7 +81,9 @@ class Defense extends Attribute {
   String toString() => 'Defense($points)';
 }
 
+/// A character trait that can increase the damage during a fight
 class Magik extends Attribute {
+  /// Initialize magik [points]
   const Magik(int points) : super(points);
 
   @override
