+++ lib/src/core/data/models/fight.dart
@@ -2,22 +2,30 @@ import 'package:fight_club/src/core/data/models/models.dart';
 import 'package:flutter/foundation.dart';
 import 'package:uuid/uuid.dart';
 
+/// Unique object that store fight history.
 @immutable
 class Fight extends Model {
+  /// A [Fight] is representing by a date and rounds history.
   Fight({
     String? id,
     required this.date,
     this.rounds = const [],
   }) : super(id ?? const Uuid().v4());
 
+  /// The date on which the fight took place.
   final DateTime date;
+
+  /// The [Round] collection.
   final List<Round> rounds;
 
+  /// Find the identifier of the last fighter who attacked and returns if it
+  /// matches given [Character]'s id.
   bool didWin(Character character) {
     if (rounds.isEmpty) return false;
     return rounds.last.attacker.id == character.id;
   }
 
+  /// Clone this object with different values.
   Fight copyWith({DateTime? date, List<Round>? rounds}) {
     return Fight(date: date ?? this.date, rounds: rounds ?? this.rounds);
   }
@@ -38,8 +46,10 @@ class Fight extends Model {
   int get hashCode => date.hashCode ^ rounds.hashCode;
 }
 
+/// Value object of [Fight] details.
 @immutable
 class FightResult {
+  /// Gives the result of a fight after it has been processed.
   const FightResult({
     required this.character,
     required this.opponent,
@@ -47,9 +57,16 @@ class FightResult {
     required this.fight,
   });
 
+  /// The character who launched the fight
   final Character character;
+
+  /// The character who has been chosen as the opponent.
   final Character opponent;
+
+  /// The resut for the character who launched the fight.
   final bool didWin;
+
+  /// The fight details.
   final Fight fight;
 
   @override
