+++ lib/src/modules/authentication/logic/auth_controller.dart
@@ -5,13 +5,15 @@ import 'package:fight_club/src/core/data/models/models.dart';
 import 'package:fight_club/src/core/storage/storage.dart';
 import 'package:fight_club/src/modules/lobby/logic/fight_observer.dart';
 
+/// maximum characters per player
 const kCharactersLengthLimit = 10;
 
 /// A class that many Widgets can interact with to read, update, or listen to
 /// [Session] changes.
 class AuthController extends AuthService implements FightObserver {
-  AuthController();
-
+  /// Adds [character] to the end of the list.
+  /// @throw [CharactersLengthLimitException] if list have reached the lenght
+  /// limit.
   void addNewCharacter(Character character) {
     if (state.characters.length >= kCharactersLengthLimit) {
       throw CharactersLengthLimitException();
@@ -20,12 +22,14 @@ class AuthController extends AuthService implements FightObserver {
     state = state.copyWith(characters: [...state.characters, character]);
   }
 
+  /// Update character where the ids match.
   void updateCharacter(Character character) {
     state = state.copyWith(
       characters: state.characters.replace(character),
     );
   }
 
+  /// Remove characters where the ids match.
   void removeCharacter(Character character) {
     state = state.copyWith(
       characters: state.characters.delete(character),
@@ -49,24 +53,27 @@ class AuthController extends AuthService implements FightObserver {
   }
 }
 
-/// A service that stores and retrieves auth info.
+/// A service that stores and retrieves auth informations from internal storage.
 abstract class AuthService extends HydratedStateNotifier<Session> {
+  /// Store an empty session by default.
   AuthService() : super(const Session());
 
+  /// Serialization and deserialization service for [Session].
   static const codec = SessionCodec();
 
   @override
   void hydrate() {/* do not persist the initial state to storage on start */}
 
-  /// Loads [Session] from local storage.
+  /// Loads [Session] from internal storage.
   @override
   Session? fromJson(Map<String, dynamic> json) => codec.fromMap(json);
 
-  /// Persists [Session] to local storage.
+  /// Persists [Session] to internal storage.
   @override
   Map<String, dynamic>? toJson(Session state) => codec.toMap(state);
 }
 
+/// Thrown when player's characters list exceed the lenght limit.
 class CharactersLengthLimitException implements Exception {
   @override
   String toString() {
