+++ lib/src/modules/lobby/logic/fight_observer.dart
@@ -1,5 +1,7 @@
 import 'package:fight_club/src/core/data/models/models.dart';
 
+/// An object that listens for fight results.
 class FightObserver {
+  /// A fight was processed.
   void didFight(Character character, Fight fight) {}
 }
