+++ lib/src/modules/lobby/logic/fight_service.dart
@@ -5,17 +5,21 @@ import 'package:fight_club/src/core/data/random_generator.dart';
 import 'package:fight_club/src/modules/lobby/logic/fight_observer.dart';
 import 'package:flutter/foundation.dart';
 
+/// Service to process fight and notify observers.
 class FightService {
+  /// initialize list of [FightObserver]
   FightService({
     this.observers = const [],
   });
 
+  /// The fights listeners.
   final List<FightObserver> observers;
 
+  /// Random integer generator.
   @visibleForTesting
   static Dice dice = const Dice();
 
-  /// Run fight in isolate.
+  /// Process new fight in isolate and notify observers.
   /// @throw FightException if both characters can't win.
   Future<Fight> launchFight(Character user, Character opponent) async {
     assertAtLeastOneCharacterCanWin(user, opponent);
@@ -33,6 +37,8 @@ class FightService {
     return fight;
   }
 
+  /// Produce rounds until one of the characters has no more health points.
+  /// The last character who attacked will be the winner.
   @visibleForTesting
   static List<Round> processFight(IsolateEntry entry) {
     final rounds = <Round>[];
@@ -58,6 +64,10 @@ class FightService {
     return rounds;
   }
 
+  /// Roll a dice with as many faces as the [attacker]'s [Attack] attribute,
+  /// and compare the result with [defender]'s [Defense] attribute to compute
+  /// the round damages.
+  /// Return a [Round] with [defender]'s [Health] updated.
   @visibleForTesting
   static Round processRound(
     int roundId,
@@ -89,6 +99,8 @@ class FightService {
     );
   }
 
+  /// Throw [FightException] to avoid infinite loop on a fight where both
+  /// character cant create damages on each other.
   void assertAtLeastOneCharacterCanWin(Character first, Character second) {
     final firstCanWin = first<Attack>().points > second<Defense>().points;
     final secondCanWin = second<Attack>().points > first<Defense>().points;
@@ -97,20 +109,28 @@ class FightService {
   }
 }
 
+/// Dice roller.
 class Dice {
+  /// Create a dice roller.
   const Dice();
 
   /// Generates a random integer between 1 ~ max (inclusive).
   int roll(int max) => Random.integer(max + 1, min: 1);
 }
 
+/// Store fighter informations.
 class IsolateEntry {
+  /// Arguments for processing a fight on isolate.
   IsolateEntry(this.characterA, this.characterB);
 
+  /// The first character to attack.
   final Character characterA;
+
+  /// The first character to defend.
   final Character characterB;
 }
 
+/// Thrown when characters cant create damages on each other.
 class FightException implements Exception {
   @override
   String toString() {
