+++ lib/src/modules/lobby/views/fight_result.dart
@@ -3,9 +3,13 @@ import 'package:fight_club/src/modules/lobby/views/layout.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+/// Process fight and show repport.
 class FightResultView extends ConsumerWidget {
+  /// Render CircularProgressIndicator during fight processing and then either
+  /// the failure reason or FightResumeLayout on success.
   const FightResultView({Key? key}) : super(key: key);
 
+  /// MaterialPageRoute that will render FightResultView
   static Route<void> route() {
     return MaterialPageRoute<void>(builder: (_) => const FightResultView());
   }
